+++ libfs/src/filesystem/file.h
@@ -37,9 +37,9 @@ struct inode *mlfs_object_create(const char *path, unsigned short mode, uint8_t 
 
 int mlfs_file_stat(struct file *f, struct stat *st);
 ssize_t mlfs_file_read(struct file *f, uint8_t *buf, size_t n);
-int mlfs_file_read_offset(struct file *f, uint8_t *bug, 
+ssize_t mlfs_file_read_offset(struct file *f, uint8_t *bug, 
 		size_t n, offset_t off);
-int mlfs_file_write(struct file *f, uint8_t *buf, offset_t offset, size_t n);
+ssize_t mlfs_file_write(struct file *f, uint8_t *buf, offset_t offset, size_t n);
 int mlfs_file_fallocate(struct file *f, offset_t offset, size_t n);
 
 #endif
